requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
rank-bm25>=0.2.2
pyspellchecker>=0.8.1
streamlit>=1.35.0
numpy>=1.24.0
